add selfcare service
