+++ apps/backoffice/.env.development
@@ -33,6 +33,7 @@ API_APIM_MOCKING=true
 
 SELFCARE_BASE_URL=http://localhost:7075
 SELFCARE_API_MOCKING=true
+SELFCARE_API_KEY=SELFCARE_API_KEY
 
 LEGACY_COSMOSDB_NAME=db
 LEGACY_COSMOSDB_URI=http://localhost:7076/
