+++ apps/backoffice/src/lib/be/institutions/business.ts
@@ -1,0 +1,8 @@
+
+
+export const getInstitutions = (selfCareUserId: string) => {
+
+};
+export const getInstitution = (institutionId: string) => {
+
+};
